refactor table head cells
