+++ src/types/index.ts
@@ -1,2 +1,3 @@
 export * from '~/types/user.types'
 export * from '~/types/common.types'
+export * from '~/types/table.interfaces'
