+++ src/types/table.interfaces.ts
@@ -1,0 +1,8 @@
+import { OperatorType } from "~/types"
+
+export interface HeadCell {
+  id: keyof OperatorType | string
+  label: string
+  isSortable: boolean
+  style?: React.CSSProperties
+}
